requirements.txt
yahoo-fantasy-sports-api==2.6.0
python-dotenv==1.0.0
